Create HousesList and House separated components
